Extracting Effective/Equivalent Refractive Index Model of a Metamaterial, Ekin Gunes Ozaktas, December 2022
This code takes in simulated data of the S parameters of a metamaterial and produces an effective or equivalent refractive index model. The complex logarithm branching problem is resolved via choosing the starting branch based on the starting derivative being minimized (predicted homogeneity at long wavelengths) and then enforcing  continuity. A comparison is made between the simulated and modeled slabs to determine the effective thickness. Plots of the final effective refractive index parameters varying with wavelength are produced, alongside graphs of the reflectance, transmittance, and absorptance of the simulated and modeled slabs.